requirements.txt
alembic==1.9.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dotenv==0.21.0
six==1.16.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
zipp==3.11.0

#the SQL files are irrelevant. 
#They are just used to seed the data 
#into the database. It’s like having 
#some data filled already into the database.
